fix/organized routes
