+++ app/about/page.tsx
@@ -1,4 +1,4 @@
-import AboutContent from "@/src/ui/modules/aboutContent/AboutContent";
+import { AboutContent } from "@/src/ui/modules/aboutContent/AboutContent";
 
 export const About = () => (
     <AboutContent />
