+++ src/ui/components/MonoText.tsx
