+++ src/ui/modules/aboutContent/AboutContent.tsx
@@ -1,7 +1,7 @@
 import Image from "next/image"
 import Link from "next/link"
 
-import Divider from "@/src/ui/repeated_components/Divider"
+import {Divider} from "@/src/ui/components/Divider"
 
 import { TARGET_BLANK, ABOUT_SECT, CERTIFIC_SECT } from "../../constants/mainconstants"
 import { PARAGRAPH_ONE, PARAGRAPH_TWO, PARAGRAPH_THREE, CV_INFO, PARAGRAPH_CERTIFIC, CERTIFICATES } from "./constants/aboutitems"
