+++ src/ui/modules/footer/components/FooterInfo.tsx
@@ -1,4 +1,4 @@
-import MonoText from "@/src/ui/repeated_components/MonoText";
+import { MonoText } from "@/src/ui/components/MonoText";
 
 import { INFO_FOOTER } from "../constants/footerconstants";
