+++ src/ui/modules/header/Header.tsx
@@ -1,10 +1,14 @@
-import MonoText from "../../repeated_components/MonoText";
-import Navigation from "./components/Navigation";
+import { MonoText } from "../../components/MonoText";
+import { Burger } from "./components/Burger/Burger";
+import { Navigation } from "./components/Navigation/Navigation";
 
 export const Header = () => {
     return (
         <MonoText>
-            <Navigation />
+            <div className="grid-gap4-M grid-cols-[1fr_auto] md:grid-cols-[auto_1fr] justify-between py-6">
+                <Burger />
+                <Navigation />
+            </div>
         </MonoText>
     )
 }
