+++ src/ui/modules/header/components/Burger/Logo.tsx
@@ -1,13 +1,15 @@
 import Image from "next/image"
 
+import { LOGO_IMG } from "../../constants/headeritems"
+
 
 export const Logo = () => {
     return (
             <Image
-                src='/Logo.png'
+                src={LOGO_IMG.src}
                 width={60}
                 height={60}
-                alt="logo"
+                alt={LOGO_IMG.alt}
                 unoptimized
                 className="md:hidden lg:block"
             />
