+++ src/ui/modules/header/components/Navigation/Navigation.tsx
@@ -1,21 +1,13 @@
-import NavList from "./NavList";
-import Button from "@/src/ui/repeated_components/Button";
-import VLine from "@/src/ui/repeated_components/VLine";
-import Social from "./SocialCircles";
-import { Logo } from "./Logo";
-import { Burger } from "./Burger";
+import { Button } from "@/src/ui/components/Button";
+import { VLine } from "@/src/ui/components/VLine";
+import { NavList } from "./NavList";
+import { SocialCircles } from "./SocialCircles";
 
-import { HEADER_BUTTON_TEXT } from "../constants/headeritems";
+import { HEADER_BUTTON_TEXT } from "../../constants/headeritems";
 
 export const Navigation = () => {
     return (
         <header>
-            <div className="grid-gap4-M grid-cols-[1fr_auto] md:grid-cols-[auto_1fr] justify-between py-6">
-                <div className="flex-justify-M w-full">
-                    <Logo />
-                    <Burger />
-                </div>
-
                 <div className="hidden md:block mx-auto w-full max-w-7xl z-100">
                     <div className="flex-center-M justify-end gap-x-8 gap-y-8">
                         <NavList />
@@ -26,11 +18,10 @@ export const Navigation = () => {
                                 className="text-whiteM bg-blackM border border-blackM py-[0.6rem] px-[1rem] text-center rounded-M md:hidden lg:block hover:bg-[transparent] hover:text-blackM"
                             />
                             <VLine />
-                            <Social />
+                            <SocialCircles />
                         </div>
                     </div>
                 </div>
-            </div>
         </header>
     )
 }
