+++ src/ui/modules/homeContent/components/SkillsHome/SvgSkills.tsx
